article-search/requirements.txt
requests
starlette
pydantic
fastapi
uvicorn
urllib3
certifi
kafka-python
elasticsearch
